+++ src/Database/QueryBuilder/PgsqlQueryBuilder.php
@@ -104,7 +104,6 @@ class PgsqlQueryBuilder extends CommonQueryBuilder implements QueryBuilderInterf
         }
         
         $columns = $table->getColumns();
-        unset($columns['id']);
         if (!empty($columns)) {
             $query = 'ALTER TABLE ' . $this->escapeString($table->getName()) . ' ';
             $columnList = [];
@@ -122,7 +121,6 @@ class PgsqlQueryBuilder extends CommonQueryBuilder implements QueryBuilderInterf
         }
         
         $primaryColumns = $table->getPrimaryColumns();
-        unset($primaryColumns['id']);
         if (!empty($primaryColumns)) {
             $queries[] = 'ALTER TABLE ' . $this->escapeString($table->getName()) . ' ADD ' . $this->primaryKeyString($table, $primaryColumns) . ';';
         }